refactor: split up sidebar
