+++ apps/web/components/modals/UserProfile/sidebar/index.tsx
@@ -1,0 +1,13 @@
+import { Tab } from '@headlessui/react'
+import SidebarButton from './button'
+
+export default function Sidebar() {
+  return (
+    <Tab.List className="w-48 flex flex-col">
+      <SidebarButton>Public profile</SidebarButton>
+      <SidebarButton>Account</SidebarButton>
+      <SidebarButton>Graphs</SidebarButton>
+      <SidebarButton>Degrees</SidebarButton>
+    </Tab.List>
+  )
+}
